requirements.txt
python-docx>=0.8.11
requests>=2.28.0
python-dotenv>=1.0.0
